tests/src/requirements.txt
paho-mqtt==2.1.0
numpy==2.3.5